+++ src/main/java/com/thecar/thecar/repositories/VeiculoRepository.java
@@ -3,6 +3,6 @@ package com.thecar.thecar.repositories;
 import com.thecar.thecar.entities.Veiculo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VeiculoRepository extends JpaRepository<Veiculo, Integer> {
+public interface VeiculoRepository extends JpaRepository<Veiculo, Long> {
 
 }
